# Fix and accelerate generated __hash__ methods