We now create a method like we do for `__init__ and` hash also the class of the
instance.

Performance improvements on my notebook:

```python
import attr


@attr.s(hash=True)
class C:
    x = attr.ib()
    y = attr.ib()
    z = attr.ib()


c = C(1, "2", 3.0)
```

attrs 16.3.:

```
➤ pyperf timeit --rigorous -s "from t import c" "hash(c)"
.........................................
Mean +- std dev: 1.81 us +- 0.06 us
```

attrs master:

```
➤ pyperf timeit --rigorous -s "from t import c" "hash(c)"
.........................................
Mean +- std dev: 474 ns +- 18 ns
```

This may also make @DRMacIver slightly happy.

Fixes #295 
Ref #261