@hynek 2017-11-26
Re the hashing of str/floats comment:

float:

```
.........................................
Mean +- std dev: 454 ns +- 19 ns
```

str:

```
.........................................
Mean +- std dev: 435 ns +- 19 ns
```

Also the performance of hashing strings seemed more stable over multiple runs.